refactor(streamwall): extract CLI parsing into cliArgs.ts
